Move instance tests to their own file
