+++ pkg/osquery/runtime/runtime_test.go
@@ -80,212 +80,6 @@ func TestMain(m *testing.M) {
 	os.Exit(retCode) //nolint:forbidigo // Fine to use os.Exit in tests
 }
 
-// getBinDir finds the directory of the currently running binary (where we will
-// look for the osquery extension)
-func getBinDir() (string, error) {
-	binPath, err := os.Executable()
-	if err != nil {
-		return "", err
-	}
-	binDir := filepath.Dir(binPath)
-	return binDir, nil
-}
-
-func TestCalculateOsqueryPaths(t *testing.T) {
-	t.Parallel()
-	binDir, err := getBinDir()
-	require.NoError(t, err)
-
-	paths, err := calculateOsqueryPaths(binDir, defaultRegistrationId, ulid.New(), osqueryOptions{})
-
-	require.NoError(t, err)
-
-	// ensure that all of our resulting artifact files are in the rootDir that we
-	// dictated
-	require.Equal(t, binDir, filepath.Dir(paths.pidfilePath))
-	require.Equal(t, binDir, filepath.Dir(paths.databasePath))
-
-	// socket path on windows includes semi-random ulid
-	if runtime.GOOS != "windows" {
-		require.Equal(t, binDir, filepath.Dir(paths.extensionSocketPath))
-	}
-
-	require.Equal(t, binDir, filepath.Dir(paths.extensionAutoloadPath))
-}
-
-func TestCreateOsqueryCommand(t *testing.T) {
-	t.Parallel()
-	paths := &osqueryFilePaths{
-		pidfilePath:           "/foo/bar/osquery-abcd.pid",
-		databasePath:          "/foo/bar/osquery.db",
-		extensionSocketPath:   "/foo/bar/osquery.sock",
-		extensionAutoloadPath: "/foo/bar/osquery.autoload",
-	}
-
-	osquerydPath := testOsqueryBinaryDirectory
-
-	osqOpts := &osqueryOptions{
-		stdout: os.Stdout,
-		stderr: os.Stderr,
-	}
-	k := typesMocks.NewKnapsack(t)
-	k.On("WatchdogEnabled").Return(true)
-	k.On("WatchdogMemoryLimitMB").Return(150)
-	k.On("WatchdogUtilizationLimitPercent").Return(20)
-	k.On("WatchdogDelaySec").Return(120)
-	k.On("OsqueryVerbose").Return(true)
-	k.On("OsqueryFlags").Return([]string{})
-	k.On("Slogger").Return(multislogger.NewNopLogger())
-
-	i := newInstance(defaultRegistrationId, k, mockServiceClient())
-	i.opts = *osqOpts
-	i.knapsack = k
-
-	cmd, err := i.createOsquerydCommand(osquerydPath, paths)
-	require.NoError(t, err)
-	require.Equal(t, os.Stderr, cmd.Stderr)
-	require.Equal(t, os.Stdout, cmd.Stdout)
-
-	k.AssertExpectations(t)
-}
-
-func TestCreateOsqueryCommandWithFlags(t *testing.T) {
-	t.Parallel()
-	osqOpts := &osqueryOptions{}
-	k := typesMocks.NewKnapsack(t)
-	k.On("WatchdogEnabled").Return(true)
-	k.On("WatchdogMemoryLimitMB").Return(150)
-	k.On("WatchdogUtilizationLimitPercent").Return(20)
-	k.On("WatchdogDelaySec").Return(120)
-	k.On("OsqueryFlags").Return([]string{"verbose=false", "windows_event_channels=foo,bar"})
-	k.On("OsqueryVerbose").Return(true)
-	k.On("Slogger").Return(multislogger.NewNopLogger())
-
-	i := newInstance(defaultRegistrationId, k, mockServiceClient())
-	i.opts = *osqOpts
-	i.knapsack = k
-
-	cmd, err := i.createOsquerydCommand(
-		testOsqueryBinaryDirectory,
-		&osqueryFilePaths{},
-	)
-	require.NoError(t, err)
-
-	// count of flags that cannot be overridden with this option
-	const nonOverridableFlagsCount = 8
-
-	// Ensure that the provided flags were placed last (so that they can override)
-	assert.Equal(
-		t,
-		[]string{"--verbose=false", "--windows_event_channels=foo,bar"},
-		cmd.Args[len(cmd.Args)-2-nonOverridableFlagsCount:len(cmd.Args)-nonOverridableFlagsCount],
-	)
-
-	k.AssertExpectations(t)
-}
-
-func TestCreateOsqueryCommand_SetsEnabledWatchdogSettingsAppropriately(t *testing.T) {
-	t.Parallel()
-
-	osqOpts := &osqueryOptions{}
-	k := typesMocks.NewKnapsack(t)
-	k.On("WatchdogEnabled").Return(true)
-	k.On("WatchdogMemoryLimitMB").Return(150)
-	k.On("WatchdogUtilizationLimitPercent").Return(20)
-	k.On("WatchdogDelaySec").Return(120)
-	k.On("Slogger").Return(multislogger.NewNopLogger())
-	k.On("OsqueryVerbose").Return(true)
-	k.On("OsqueryFlags").Return([]string{})
-
-	i := newInstance(defaultRegistrationId, k, mockServiceClient())
-	i.opts = *osqOpts
-	i.knapsack = k
-
-	cmd, err := i.createOsquerydCommand(
-		testOsqueryBinaryDirectory,
-		&osqueryFilePaths{},
-	)
-	require.NoError(t, err)
-
-	watchdogMemoryLimitMBFound := false
-	watchdogUtilizationLimitPercentFound := false
-	watchdogDelaySecFound := false
-	for _, a := range cmd.Args {
-		if strings.Contains(a, "disable_watchdog") {
-			t.Error("disable_watchdog flag set")
-			t.FailNow()
-		}
-
-		if a == "--watchdog_memory_limit=150" {
-			watchdogMemoryLimitMBFound = true
-			continue
-		}
-
-		if a == "--watchdog_utilization_limit=20" {
-			watchdogUtilizationLimitPercentFound = true
-			continue
-		}
-
-		if a == "--watchdog_delay=120" {
-			watchdogDelaySecFound = true
-			continue
-		}
-	}
-
-	require.True(t, watchdogMemoryLimitMBFound, "watchdog memory limit not set")
-	require.True(t, watchdogUtilizationLimitPercentFound, "watchdog CPU limit not set")
-	require.True(t, watchdogDelaySecFound, "watchdog delay sec not set")
-
-	k.AssertExpectations(t)
-}
-
-func TestCreateOsqueryCommand_SetsDisabledWatchdogSettingsAppropriately(t *testing.T) {
-	t.Parallel()
-
-	osqOpts := &osqueryOptions{}
-	k := typesMocks.NewKnapsack(t)
-	k.On("WatchdogEnabled").Return(false)
-	k.On("Slogger").Return(multislogger.NewNopLogger())
-	k.On("OsqueryVerbose").Return(true)
-	k.On("OsqueryFlags").Return([]string{})
-
-	i := newInstance(defaultRegistrationId, k, mockServiceClient())
-	i.opts = *osqOpts
-	i.knapsack = k
-
-	cmd, err := i.createOsquerydCommand(
-		testOsqueryBinaryDirectory,
-		&osqueryFilePaths{},
-	)
-	require.NoError(t, err)
-
-	disableWatchdogFound := false
-	for _, a := range cmd.Args {
-		if strings.Contains(a, "watchdog_memory_limit") {
-			t.Error("watchdog_memory_limit flag set")
-			t.FailNow()
-		}
-
-		if strings.Contains(a, "watchdog_utilization_limit") {
-			t.Error("watchdog_utilization_limit flag set")
-			t.FailNow()
-		}
-
-		if strings.Contains(a, "watchdog_delay") {
-			t.Error("watchdog_delay flag set")
-			t.FailNow()
-		}
-
-		if strings.Contains(a, "disable_watchdog") {
-			disableWatchdogFound = true
-		}
-	}
-
-	require.True(t, disableWatchdogFound, "watchdog disabled not set")
-
-	k.AssertExpectations(t)
-}
-
 // downloadOsqueryInBinDir downloads osqueryd. This allows the test
 // suite to run on hosts lacking osqueryd. We could consider moving this into a deps step.
 func downloadOsqueryInBinDir(binDirectory string) error {
@@ -496,7 +290,7 @@ func waitHealthy(t *testing.T, runner *Runner, logBytes *threadsafebuffer.Thread
 		if runner.instances[defaultRegistrationId] == nil {
 			return errors.New("default instance does not exist yet")
 		}
-		if runner.instances[defaultRegistrationId].stats.ConnectTime == "" {
+		if runner.instances[defaultRegistrationId].stats == nil || runner.instances[defaultRegistrationId].stats.ConnectTime == "" {
 			return errors.New("no connect time set yet")
 		}
